CRUD supplier

+++ src/main/java/com/sep490/bads/distributionsystem/dto/SupplierStatisticsDto.java
@@ -9,6 +9,6 @@ import lombok.*;
 @Builder
 public class SupplierStatisticsDto {
     private Long totalOrders;  // Tổng số đơn mua hàng
-    private Double totalDebt;  // Tổng công nợ
-    private Double totalPaid;  // Tổng tiền đã thanh toán
+    private Long totalDebt;  // Tổng công nợ
+    private Long totalPaid;  // Tổng tiền đã thanh toán
 }
